divides DataModule: NetworkModule, NetworkCommonModule
