+++ app/src/main/java/com/example/compmoviedb/di/NetworkModule.kt
@@ -9,22 +9,11 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
 
-@InstallIn(SingletonComponent::class)
 @Module
-class DataModule {
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(): MovieDBApiInterface =
-        Retrofit.Builder()
-            .baseUrl(MovieDBApiInterface.BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(MovieDBApiInterface::class.java)
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
 
     @Provides
     @Singleton
